Refactor: "Header 이름 변경, 파일 합치기"
Navigation으로 변경
